+++ database/install/myems_system_db.sql
@@ -319,6 +319,144 @@ CREATE TABLE IF NOT EXISTS `myems_system_db`.`tbl_energy_storage_power_stations`
   PRIMARY KEY (`id`));
 CREATE INDEX `tbl_energy_storage_power_stations_index_1` ON  `myems_system_db`.`tbl_energy_storage_power_stations` (`name`);
 
+
+-- ---------------------------------------------------------------------------------------------------------------------
+-- Table `myems_system_db`.`tbl_energy_storage_power_stations_batteries`
+-- ---------------------------------------------------------------------------------------------------------------------
+DROP TABLE IF EXISTS `myems_system_db`.`tbl_energy_storage_power_stations_batteries` ;
+
+CREATE TABLE IF NOT EXISTS `myems_system_db`.`tbl_energy_storage_power_stations_batteries` (
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
+  `name` VARCHAR(255) NOT NULL,
+  `uuid` CHAR(36) NOT NULL,
+  `energy_storage_power_station_id` BIGINT NOT NULL,
+  `battery_state_point_id` BIGINT NOT NULL,
+  `soc_point_id` BIGINT NOT NULL,
+  `power_point_id` BIGINT NOT NULL,
+  `charge_meter_id` BIGINT NOT NULL,
+  `discharge_meter_id` BIGINT NOT NULL,
+  `capacity` DECIMAL(18, 3) NOT NULL,
+  `nominal_voltage` DECIMAL(18, 3) NOT NULL,
+  PRIMARY KEY (`id`));
+CREATE INDEX `tbl_energy_storage_power_stations_batteries_index_1` ON  `myems_system_db`.`tbl_energy_storage_power_stations_batteries` (`name`);
+
+-- ---------------------------------------------------------------------------------------------------------------------
+-- Table `myems_system_db`.`tbl_energy_storage_power_stations_commands`
+-- ---------------------------------------------------------------------------------------------------------------------
+DROP TABLE IF EXISTS `myems_system_db`.`tbl_energy_storage_power_stations_commands` ;
+
+CREATE TABLE IF NOT EXISTS `myems_system_db`.`tbl_energy_storage_power_stations_commands` (
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
+  `energy_storage_power_station_id` BIGINT NOT NULL,
+  `command_id` BIGINT NOT NULL,
+  PRIMARY KEY (`id`));
+CREATE INDEX `tbl_energy_storage_power_stations_commands_index_1` ON  `myems_system_db`.`tbl_energy_storage_power_stations_commands`   (`energy_storage_power_station_id`);
+
+-- ---------------------------------------------------------------------------------------------------------------------
+-- Table `myems_system_db`.`tbl_energy_storage_power_stations_power_conversion_systems`
+-- ---------------------------------------------------------------------------------------------------------------------
+DROP TABLE IF EXISTS `myems_system_db`.`tbl_energy_storage_power_stations_power_conversion_systems` ;
+
+CREATE TABLE IF NOT EXISTS `myems_system_db`.`tbl_energy_storage_power_stations_power_conversion_systems` (
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
+  `name` VARCHAR(255) NOT NULL,
+  `uuid` CHAR(36) NOT NULL,
+  `energy_storage_power_station_id` BIGINT NOT NULL,
+  `run_state_point_id` BIGINT NOT NULL,
+  `capacity` DECIMAL(18, 3) NOT NULL,
+  `charge_start_time1_point_id` BIGINT NOT NULL,
+  `charge_end_time1_point_id` BIGINT NOT NULL,
+  `charge_start_time2_point_id` BIGINT NOT NULL,
+  `charge_end_time2_point_id` BIGINT NOT NULL,
+  `charge_start_time3_point_id` BIGINT NOT NULL,
+  `charge_end_time3_point_id` BIGINT NOT NULL,
+  `charge_start_time4_point_id` BIGINT NOT NULL,
+  `charge_end_time4_point_id` BIGINT NOT NULL,
+  `discharge_start_time1_point_id` BIGINT NOT NULL,
+  `discharge_end_time1_point_id` BIGINT NOT NULL,
+  `discharge_start_time2_point_id` BIGINT NOT NULL,
+  `discharge_end_time2_point_id` BIGINT NOT NULL,
+  `discharge_start_time3_point_id` BIGINT NOT NULL,
+  `discharge_end_time3_point_id` BIGINT NOT NULL,
+  `discharge_start_time4_point_id` BIGINT NOT NULL,
+  `discharge_end_time4_point_id` BIGINT NOT NULL,
+  `charge_start_time1_command_id` BIGINT NOT NULL,
+  `charge_end_time1_command_id` BIGINT NOT NULL,
+  `charge_start_time2_command_id` BIGINT NOT NULL,
+  `charge_end_time2_command_id` BIGINT NOT NULL,
+  `charge_start_time3_command_id` BIGINT NOT NULL,
+  `charge_end_time3_command_id` BIGINT NOT NULL,
+  `charge_start_time4_command_id` BIGINT NOT NULL,
+  `charge_end_time4_command_id` BIGINT NOT NULL,
+  `discharge_start_time1_command_id` BIGINT NOT NULL,
+  `discharge_end_time1_command_id` BIGINT NOT NULL,
+  `discharge_start_time2_command_id` BIGINT NOT NULL,
+  `discharge_end_time2_command_id` BIGINT NOT NULL,
+  `discharge_start_time3_command_id` BIGINT NOT NULL,
+  `discharge_end_time3_command_id` BIGINT NOT NULL,
+  `discharge_start_time4_command_id` BIGINT NOT NULL,
+  `discharge_end_time4_command_id` BIGINT NOT NULL,
+  PRIMARY KEY (`id`));
+CREATE INDEX `tbl_energy_storage_power_stations_pcs_index_1` ON  `myems_system_db`.`tbl_energy_storage_power_stations_power_conversion_systems` (`name`);
+
+-- ---------------------------------------------------------------------------------------------------------------------
+-- Table `myems_system_db`.`tbl_energy_storage_power_stations_grids`
+-- ---------------------------------------------------------------------------------------------------------------------
+DROP TABLE IF EXISTS `myems_system_db`.`tbl_energy_storage_power_stations_grids` ;
+
+CREATE TABLE IF NOT EXISTS `myems_system_db`.`tbl_energy_storage_power_stations_grids` (
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
+  `name` VARCHAR(255) NOT NULL,
+  `uuid` CHAR(36) NOT NULL,
+  `energy_storage_power_station_id` BIGINT NOT NULL,
+  `power_point_id` BIGINT NOT NULL,
+  `buy_meter_id` BIGINT NOT NULL,
+  `sell_meter_id` BIGINT NOT NULL,
+  `capacity` DECIMAL(18, 3) NOT NULL,
+  PRIMARY KEY (`id`));
+CREATE INDEX `tbl_energy_storage_power_stations_grids_index_1` ON  `myems_system_db`.`tbl_energy_storage_power_stations_grids` (`name`);
+
+-- ---------------------------------------------------------------------------------------------------------------------
+-- Table `myems_system_db`.`tbl_energy_storage_power_stations_loads`
+-- ---------------------------------------------------------------------------------------------------------------------
+DROP TABLE IF EXISTS `myems_system_db`.`tbl_energy_storage_power_stations_loads` ;
+
+CREATE TABLE IF NOT EXISTS `myems_system_db`.`tbl_energy_storage_power_stations_loads` (
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
+  `name` VARCHAR(255) NOT NULL,
+  `uuid` CHAR(36) NOT NULL,
+  `energy_storage_power_station_id` BIGINT NOT NULL,
+  `power_point_id` BIGINT NOT NULL,
+  `meter_id` BIGINT NOT NULL,
+  `capacity` DECIMAL(18, 3) NOT NULL,
+  PRIMARY KEY (`id`));
+CREATE INDEX `tbl_energy_storage_power_stations_grids_index_1` ON  `myems_system_db`.`tbl_energy_storage_power_stations_loads` (`name`);
+
+-- ---------------------------------------------------------------------------------------------------------------------
+-- Table `myems_system_db`.`tbl_energy_storage_power_stations_sensors`
+-- ---------------------------------------------------------------------------------------------------------------------
+DROP TABLE IF EXISTS `myems_system_db`.`tbl_energy_storage_power_stations_sensors` ;
+
+CREATE TABLE IF NOT EXISTS `myems_system_db`.`tbl_energy_storage_power_stations_sensors` (
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
+  `energy_storage_power_station_id` BIGINT NOT NULL,
+  `sensor_id` BIGINT NOT NULL,
+  PRIMARY KEY (`id`));
+CREATE INDEX `tbl_energy_storage_power_stations_sensors_index_1` ON  `myems_system_db`.`tbl_energy_storage_power_stations_sensors` (`energy_storage_power_station_id`);
+
+-- ---------------------------------------------------------------------------------------------------------------------
+-- Table `myems_system_db`.`tbl_energy_storage_power_stations_users`
+-- ---------------------------------------------------------------------------------------------------------------------
+DROP TABLE IF EXISTS `myems_system_db`.`tbl_energy_storage_power_stations_users` ;
+
+CREATE TABLE IF NOT EXISTS `myems_system_db`.`tbl_energy_storage_power_stations_users` (
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
+  `energy_storage_power_station_id` BIGINT NOT NULL,
+  `user_id` BIGINT NOT NULL COMMENT 'primary key in myems_user_db.tbl_users',
+  PRIMARY KEY (`id`));
+CREATE INDEX `tbl_energy_storage_power_stations_users_index_1` ON  `myems_system_db`.`tbl_energy_storage_power_stations_users` (`energy_storage_power_station_id`);
+
+
 -- ---------------------------------------------------------------------------------------------------------------------
 -- Table `myems_system_db`.`tbl_equipments`
 -- ---------------------------------------------------------------------------------------------------------------------